lines cleaned i know you didnt ask but daniel died for our sins god bless
